unique module names via sha1 of info
